finish task 1

+++ src/DataTypes.java
@@ -1,7 +1,6 @@
 public class DataTypes {
-    // TODO TASK 1: fix this code so that it passes the test in DataTypesTest.java
     public static long sum(int[] numbers) {
-        int s = 0; // variable to accumulate the sum in!
+        long s = 0; // variable to accumulate the sum in!
         // below is a "foreach" loop which iterates through numbers
         for (int x : numbers) {
             s += x;
@@ -9,3 +8,4 @@ public class DataTypes {
         return s;
     }
 }
+// solution: change int s = 0 to long s = 0 bc the sum is large
